move files from old Common folder to new FitOnFhir.Common project
