+++ src/FitOnFhir.Common/Handlers/UnknownDataImportHandler.cs
@@ -3,15 +3,15 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System.Threading;
-using GoogleFitOnFhir.Models;
+using FitOnFhir.Common.Requests;
 
-namespace GoogleFitOnFhir.Common
+namespace FitOnFhir.Common.Handlers
 {
-    public class ImportRequest
+    public class UnknownDataImportHandler : UnknownOperationHandlerBase<ImportRequest, Task>
     {
-        public QueueMessage Message { get; set; }
-
-        public CancellationToken Token { get; set; }
+        public override Task Evaluate(ImportRequest request)
+        {
+            return Task.CompletedTask;
+        }
     }
 }
