+++ src/FitOnFhir.Common/Models/QueueMessage.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace GoogleFitOnFhir.Models
+namespace FitOnFhir.Common.Models
 {
     public class QueueMessage
     {
