+++ src/GoogleFitOnFhir.Identity/Startup.cs
@@ -7,9 +7,11 @@ using System;
 using System.Threading.Tasks;
 using Azure.Identity;
 using Azure.Security.KeyVault.Secrets;
+using FitOnFhir.Common.ExtensionMethods;
+using FitOnFhir.Common.Handlers;
+using FitOnFhir.Common.Requests;
 using GoogleFitOnFhir.Clients.GoogleFit;
 using GoogleFitOnFhir.Clients.GoogleFit.Handlers;
-using GoogleFitOnFhir.Common;
 using GoogleFitOnFhir.Persistence;
 using GoogleFitOnFhir.Repositories;
 using GoogleFitOnFhir.Services;
