+++ src/GoogleFitOnFhir.PublishData/Startup.cs
@@ -8,12 +8,14 @@ using System.Threading.Tasks;
 using Azure.Identity;
 using Azure.Messaging.EventHubs.Producer;
 using Azure.Security.KeyVault.Secrets;
+using FitOnFhir.Common;
+using FitOnFhir.Common.ExtensionMethods;
+using FitOnFhir.Common.Handlers;
+using FitOnFhir.Common.Requests;
 using GoogleFitOnFhir.Clients.GoogleFit;
 using GoogleFitOnFhir.Clients.GoogleFit.Config;
 using GoogleFitOnFhir.Clients.GoogleFit.Handlers;
 using GoogleFitOnFhir.Clients.GoogleFit.Telemetry;
-using GoogleFitOnFhir.Common;
-using GoogleFitOnFhir.Identity;
 using GoogleFitOnFhir.Persistence;
 using GoogleFitOnFhir.Repositories;
 using GoogleFitOnFhir.Services;
