+++ src/GoogleFitOnFhir.UnitTests/DependencyInjectionExtensionsTests.cs
@@ -5,8 +5,10 @@
 
 using System;
 using System.Threading.Tasks;
+using FitOnFhir.Common.ExtensionMethods;
+using FitOnFhir.Common.Handlers;
+using FitOnFhir.Common.Requests;
 using GoogleFitOnFhir.Clients.GoogleFit.Handlers;
-using GoogleFitOnFhir.Common;
 using GoogleFitOnFhir.Identity;
 using GoogleFitOnFhir.UnitTests.Mocks;
 using Microsoft.AspNetCore.Mvc;
