+++ src/GoogleFitOnFhir.UnitTests/GoogleFitDataImportHandlerTests.cs
@@ -6,9 +6,10 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using FitOnFhir.Common.Models;
+using FitOnFhir.Common.Requests;
 using GoogleFitOnFhir.Clients.GoogleFit.Handlers;
 using GoogleFitOnFhir.Common;
-using GoogleFitOnFhir.Models;
 using GoogleFitOnFhir.Services;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
