+++ src/GoogleFitOnFhir.UnitTests/GoogleFitDataImporterTests.cs
@@ -7,11 +7,11 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using FitOnFhir.Common.Models;
 using GoogleFitOnFhir.Clients.GoogleFit;
 using GoogleFitOnFhir.Clients.GoogleFit.Models;
 using GoogleFitOnFhir.Clients.GoogleFit.Responses;
 using GoogleFitOnFhir.Common;
-using GoogleFitOnFhir.Models;
 using GoogleFitOnFhir.Repositories;
 using GoogleFitOnFhir.Services;
 using GoogleFitOnFhir.UnitTests.Mocks;
