+++ src/GoogleFitOnFhir/Clients/GoogleFit/Handlers/GoogleFitDataImportHandler.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Threading.Tasks;
 using EnsureThat;
+using FitOnFhir.Common.Requests;
 using GoogleFitOnFhir.Common;
 using GoogleFitOnFhir.Services;
 using Microsoft.Extensions.Logging;
