+++ src/GoogleFitOnFhir/Common/Constants.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using GoogleFitOnFhir.Models;
+using FitOnFhir.Common.Models;
 
 namespace GoogleFitOnFhir.Common
 {
