+++ src/GoogleFitOnFhir/Services/GoogleFitImportService.cs
@@ -10,12 +10,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Messaging.EventHubs.Producer;
 using EnsureThat;
+using FitOnFhir.Common.Exceptions;
+using FitOnFhir.Common.Models;
 using GoogleFitOnFhir.Clients.GoogleFit;
 using GoogleFitOnFhir.Clients.GoogleFit.Config;
 using GoogleFitOnFhir.Clients.GoogleFit.Models;
 using GoogleFitOnFhir.Clients.GoogleFit.Responses;
-using GoogleFitOnFhir.Common;
-using GoogleFitOnFhir.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Health.Common.Service;
 using Microsoft.Health.Logging.Telemetry;
