+++ src/GoogleFitOnFhir/Services/RoutingService.cs
@@ -7,7 +7,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using EnsureThat;
-using GoogleFitOnFhir.Common;
+using FitOnFhir.Common.Requests;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
